app/integrations/yahoo_finance.py
"""
Yahoo Finance integration
"""

import requests
from typing import Dict, List, Optional


class YahooFinance:
    """Yahoo Finance API wrapper"""
    
    BASE_URL = "https://query1.finance.yahoo.com"
    
    def __init__(self):
        """Initialize Yahoo Finance client"""
        self.session = requests.Session()
    
    def get_historical_data(self, ticker: str, period: str = "1y") -> List[Dict]:
        """Get historical stock data"""
        return []
    
    def get_quote(self, ticker: str) -> Dict:
        """Get current stock quote"""
        return {
            "ticker": ticker,
            "price": 0.0,
            "change": 0.0
        }
